pyproject.toml
[project]
name = "agronexus"
version = "1.0.0"
description = "AgroNexus — AI-powered precision agriculture OpenEnv environment"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn==0.30.6",
    "pydantic==2.9.2",
    "openai==1.51.0",
    "pyyaml==6.0.2",
    "requests==2.32.3",
]

[project.scripts]
server = "uvicorn:main"

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
